+++ packages/form/src/field/CheckboxField.tsx
@@ -17,7 +17,7 @@ import {
 import { useEventCallback } from "usehooks-ts";
 
 import { CheckBoxProps, Checkbox } from "../component/Checkbox";
-import { useControlledField } from "../form/useField";
+import { useControlledField } from "../form/_internal/useField";
 import { useCustomEventDispatcher } from "../hook/useCustomEventDispatcher";
 import { CheckboxFieldGroupContext } from "./CheckboxFieldGroup";
 import {
