+++ packages/form/src/field/DatePickerField.tsx
@@ -16,8 +16,8 @@ import {
 import { useEventCallback } from "usehooks-ts";
 
 import { DatePicker, DatePickerProps } from "../component/DatePicker";
+import { useControlledField } from "../form/_internal/useField";
 import { FormAdapterContext } from "../form/FormAdapter";
-import { useControlledField } from "../form/useField";
 import { useCustomEventDispatcher } from "../hook/useCustomEventDispatcher";
 import {
   FieldEventDetail,
