+++ packages/form/src/field/DateTimePickerField.tsx
@@ -19,8 +19,8 @@ import {
   DateTimePicker,
   DateTimePickerProps,
 } from "../component/DateTimePicker";
+import { useControlledField } from "../form/_internal/useField";
 import { FormAdapterContext } from "../form/FormAdapter";
-import { useControlledField } from "../form/useField";
 import { useCustomEventDispatcher } from "../hook/useCustomEventDispatcher";
 import {
   FieldEventDetail,
