+++ packages/form/src/field/MultiAutoCompleteField.tsx
@@ -17,7 +17,7 @@ import {
   MultiAutoComplete,
   MultiAutoCompleteProps,
 } from "../component/MultiAutoComplete";
-import { useControlledField } from "../form/useField";
+import { useControlledField } from "../form/_internal/useField";
 import { useCustomEventDispatcher } from "../hook/useCustomEventDispatcher";
 import {
   FieldEventDetail,
