+++ packages/form/src/field/autocomplete/AutoCompleteField.tsx
@@ -14,7 +14,7 @@ import {
   AutoCompleteProps,
 } from "../../component/autocomplete/AutoComplete";
 import { DefaultAutoCompleteOption } from "../../component/AutoCompleteModel";
-import { useControlledField } from "../../form/useField";
+import { useControlledField } from "../../form/_internal/useField";
 import { useCustomEventDispatcher } from "../../hook/useCustomEventDispatcher";
 import {
   FieldEventDetail,
