+++ packages/form/src/field/autocomplete/CreatableAutoCompleteField.tsx
@@ -14,7 +14,7 @@ import {
   CreatableAutoCompleteProps,
 } from "../../component/autocomplete/CreatableAutoComplete";
 import { DefaultAutoCompleteOption } from "../../component/AutoCompleteModel";
-import { useControlledField } from "../../form/useField";
+import { useControlledField } from "../../form/_internal/useField";
 import { useCustomEventDispatcher } from "../../hook/useCustomEventDispatcher";
 import {
   FieldEventDetail,
