+++ packages/form/src/form/_internal/useFormActions.tsx
@@ -1,7 +1,7 @@
 import { useContext } from "react";
 
-import { FormActions } from "../field/types";
-import { FormActionContext } from "./context/FormActionContext";
+import { FormActions } from "../../field/types";
+import { FormActionContext } from "../context/FormActionContext";
 
 export function useFormActions<
   FormValues extends {}
